+++ app/app/Http/Requests/UpdateThemeSettingsRequest.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class UpdateThemeSettingsRequest extends FormRequest
+{
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * @return array<string, mixed>
+     */
+    public function rules(): array
+    {
+        return [
+            'primary_color' => ['required', 'regex:/^#[0-9a-fA-F]{6}$/'],
+            'text_color' => ['required', 'regex:/^#[0-9a-fA-F]{6}$/'],
+        ];
+    }
+}